requirements.txt
python-dotenv
Django==5.0
psycopg2-binary
mysqlclient
djangorestframework
openpyxl